data/requirements.txt
ccxt>=1.55.52
pandas>=1.4.1
psycopg2-binary>=2.9.3
